refactor(agent): split model handlers into dedicated files
* Split model_handler_others into model_handler_google & model_handler_openrouter
* Fix OpenRouter model detection in execute.py
* Add model handler debug logging
